promoted code into headmap.py, added get_heatmap
